Remove patterns
