+++ packages/_internal/lib/models.dart
@@ -42,6 +42,7 @@ class DeepCloneableProperty with _$DeepCloneableProperty {
 class ConstructorDetails with _$ConstructorDetails {
   factory ConstructorDetails({
     required String name,
+    required bool isExternal,
     required String unionValue,
     required bool isConst,
     required String redirectedName,
@@ -65,24 +66,6 @@ class ConstructorDetails with _$ConstructorDetails {
 }
 
 @freezed
-class MapConfig with _$MapConfig {
-  factory MapConfig({
-    required bool map,
-    required bool mapOrNull,
-    required bool maybeMap,
-  }) = _MapConfig;
-}
-
-@freezed
-class WhenConfig with _$WhenConfig {
-  factory WhenConfig({
-    required bool when,
-    required bool whenOrNull,
-    required bool maybeWhen,
-  }) = _WhenConfig;
-}
-
-@freezed
 class Data with _$Data {
   @Assert('constructors.isNotEmpty')
   factory Data({
@@ -91,8 +74,6 @@ class Data with _$Data {
     required bool generateCopyWith,
     required bool generateEqual,
     required bool generateToString,
-    required MapConfig map,
-    required WhenConfig when,
     required bool generateFromJson,
     required bool generateToJson,
     required bool makeCollectionsImmutable,
