+++ packages/freezed/README.md
@@ -6,11 +6,7 @@
 
 [<img src="https://raw.githubusercontent.com/rrousselGit/provider/master/resources/flutter_favorite.png" width="200" />](https://flutter.dev/docs/development/packages-and-plugins/favorites)
 
-Welcome to [Freezed], yet another code generator for data classes, unions, and cloning.
-
-> [!NOTE]
-> An early preview of Freezed using macro is available.
-> To learn more, go to <https://github.com/rrousselGit/freezed/tree/macros>.
+Welcome to [Freezed], yet another code generator for data classes, tagged unions, nested classes and cloning.
 
 # Motivation
 
@@ -55,9 +51,6 @@ to focus on the definition of your model.
   - [Union types](#union-types)
     - [Shared properties](#shared-properties)
     - [Using pattern matching to read non-shared properties](#using-pattern-matching-to-read-non-shared-properties)
-      - [(Legacy) Pattern matching utilities](#legacy-pattern-matching-utilities)
-        - [When](#when)
-        - [Map](#map)
   - [Configurations](#configurations)
     - [Changing the behavior for a specific model](#changing-the-behavior-for-a-specific-model)
     - [Changing the behavior for the entire project](#changing-the-behavior-for-the-entire-project)
@@ -423,7 +416,7 @@ class Person with _$Person {
 
 then it will fail with the error `The non-abstract class _$_Person is missing implementations for these members:`.
 
-For that to work, we need to define a private empty constructor. That will enable the generated code to *extend/subclass* our class, instead of *implementing* it (which is the default, and only inherits type, and not properties or methods):
+For that to work, we need to define a private empty constructor. That will enable the generated code to _extend/subclass_ our class, instead of _implementing_ it (which is the default, and only inherits type, and not properties or methods):
 
 ```dart
 @freezed
@@ -1009,124 +1002,6 @@ switch (example) {
 
 If using Dart 2, you can use also use the [legacy pattern matching utilities](#legacy-pattern-matching-utilities) generated by Freezed to inspect the content of our object.
 You could also using `is`/`as` to cast an `Example` variable into either a `Person` or a `City`, but this is heavily discouraged. Use one of the other two options.
-
-#### (Legacy) Pattern matching utilities
-
-> [!WARNING]
-> As of Dart 3, Dart now has built-in pattern-matching using sealed classes.
-> As such, you no-longer need to rely on Freezed's generated methods for pattern matching.
-> Instead of using `when`/`map`, use the official Dart syntax.
->
-> The references to `when`/`map` are kept for users who have yet to
-> migrate to Dart 3.
-> But in the long term, you should stop relying on them and migrate to `switch` expressions.
-
-##### When
-
-The [when] method is the equivalent to pattern matching with destructing.  
-The prototype of the method depends on the constructors defined.
-
-For example, with:
-
-```dart
-@freezed
-sealed class Union with _$Union {
-  const factory Union(int value) = Data;
-  const factory Union.loading() = Loading;
-  const factory Union.error([String? message]) = ErrorDetails;
-}
-```
-
-Then [when] will be:
-
-```dart
-var union = Union(42);
-
-print(
-  union.when(
-    (int value) => 'Data $value',
-    loading: () => 'loading',
-    error: (String? message) => 'Error: $message',
-  ),
-); // Data 42
-```
-
-Whereas if we defined:
-
-```dart
-@freezed
-sealed class Model with _$Model {
-  factory Model.first(String a) = First;
-  factory Model.second(int b, bool c) = Second;
-}
-```
-
-Then [when] will be:
-
-```dart
-var model = Model.first('42');
-
-print(
-  model.when(
-    first: (String a) => 'first $a',
-    second: (int b, bool c) => 'second $b $c'
-  ),
-); // first 42
-```
-
-Notice how each callback matches with a constructor's name and prototype.
-
-##### Map
-
-The [map] methods are equivalent to [when], but **without** destructuring.
-
-Consider this class:
-
-```dart
-@freezed
-sealed class Model with _$Model {
-  factory Model.first(String a) = First;
-  factory Model.second(int b, bool c) = Second;
-}
-```
-
-With such class, while [when] will be:
-
-```dart
-var model = Model.first('42');
-
-print(
-  model.when(
-    first: (String a) => 'first $a',
-    second: (int b, bool c) => 'second $b $c'
-  ),
-); // first 42
-```
-
-[map] will instead be:
-
-```dart
-var model = Model.first('42');
-
-print(
-  model.map(
-    first: (First value) => 'first ${value.a}',
-    second: (Second value) => 'second ${value.b} ${value.c}'
-  ),
-); // first 42
-```
-
-This can be useful if you want to do complex operations, like [copyWith]/`toString` for example:
-
-```dart
-var model = Model.second(42, false)
-print(
-  model.map(
-    first: (value) => value,
-    second: (value) => value.copyWith(c: true),
-  )
-); // Model.second(b: 42, c: true)
-```
 
 ## Configurations
