+++ packages/freezed/lib/src/templates/abstract_template.dart
@@ -30,12 +30,6 @@ class Abstract {
 mixin _\$${data.name.public}${data.genericsDefinitionTemplate} {
 
 $abstractProperties
-$_when
-$_whenOrNull
-$_maybeWhen
-$_map
-$_mapOrNull
-$_maybeMap
 $_toJson
 ${copyWith?.abstractCopyWithGetter ?? ''}
 }
@@ -53,35 +47,5 @@ ${copyWith?.commonConcreteImpl ?? ''}
     return '/// Serializes this ${data.name} to a JSON map.\n'
         'Map<String, dynamic> toJson($_toJsonParams)'
         ' => throw $privConstUsedErrorVarName;';
-  }
-
-  String get _when {
-    if (!data.when.when) return '';
-    return '${whenPrototype(data.constructors)} => throw $privConstUsedErrorVarName;';
-  }
-
-  String get _whenOrNull {
-    if (!data.when.whenOrNull) return '';
-    return '${whenOrNullPrototype(data.constructors)} => throw $privConstUsedErrorVarName;';
-  }
-
-  String get _maybeWhen {
-    if (!data.when.maybeWhen) return '';
-    return '${maybeWhenPrototype(data.constructors)} => throw $privConstUsedErrorVarName;';
-  }
-
-  String get _map {
-    if (!data.map.map) return '';
-    return '${mapPrototype(data.constructors, data.genericsParameterTemplate)} => throw $privConstUsedErrorVarName;';
-  }
-
-  String get _mapOrNull {
-    if (!data.map.mapOrNull) return '';
-    return '${mapOrNullPrototype(data.constructors, data.genericsParameterTemplate)} => throw $privConstUsedErrorVarName;';
-  }
-
-  String get _maybeMap {
-    if (!data.map.maybeMap) return '';
-    return '${maybeMapPrototype(data.constructors, data.genericsParameterTemplate)} => throw $privConstUsedErrorVarName;';
   }
 }
