+++ packages/freezed/test/private_constructor_test.dart
@@ -6,14 +6,7 @@ void main() {
   test("mixin's default methods throw", () {
     final c = PrivateConstructor.illegal();
     expect(c, isA<PrivateConstructor>());
-    expect(() => c.map((value) => null, union1: (_) => null),
-        throwsUnsupportedError);
-    expect(() => c.maybeMap((value) => null, orElse: () {}),
-        throwsUnsupportedError);
-    expect(() => c.when((value) => null, union1: (_, __) => null),
-        throwsUnsupportedError);
-    expect(() => c.maybeWhen((value) => null, orElse: () => null),
-        throwsUnsupportedError);
+
     expect(() => c..name, throwsUnsupportedError);
   });
 }
