+++ packages/freezed/test/single_class_constructor_test.dart
@@ -429,42 +429,6 @@ void main() {
     expect('${SingleNamedCtor.named(42)}', 'SingleNamedCtor.named(a: 42)');
   });
 
-  test('single-case union does have map', () async {
-    expect(
-      SingleNamedCtor.named(42).map(
-        named: (WhateverSingleNamedCtor value) => '${value.a}',
-      ),
-      '42',
-    );
-  });
-
-  test('single-case union does have maybeMap', () async {
-    expect(
-      SingleNamedCtor.named(42).maybeMap(
-        named: (WhateverSingleNamedCtor value) => '${value.a}',
-        orElse: () => throw Exception('orElse called'),
-      ),
-      '42',
-    );
-  });
-
-  test('single-case union does have when', () async {
-    expect(
-      SingleNamedCtor.named(42).when(named: (int value) => '$value'),
-      '42',
-    );
-  });
-
-  test('single-case union does have maybeWhen', () async {
-    expect(
-      SingleNamedCtor.named(42).maybeWhen(
-        named: (int value) => '$value',
-        orElse: () => throw Exception('orElse called'),
-      ),
-      '42',
-    );
-  });
-
   test('can be created as const', () {
     expect(identical(const MyClass(a: '42'), const MyClass(a: '42')), isTrue);
   });
