+++ packages/freezed_annotation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased 3.0.0
+
+- **Breaking** removed `when`/`map` related options
+
 ## 2.4.4 - 2024-07-15
 
 - Require json_annotation ^4.8.0
